Site Updates
Fixed the navbar toggle when screen size reduced. Added jquery and popper.js to complete the bootstrap package for react. went back and changed div with class -- to className. Lastly added a redirect var to the react state for Home.js and profile.js that will take user back to login page if they do not have token in local storage.